Added POST endoints and DTOs

+++ src/main/java/com/malamute/dublin/entities/ExerciseRecord.java
@@ -1,9 +1,7 @@
 package com.malamute.dublin.entities;
 
-import java.math.BigInteger;
-
 public class ExerciseRecord {
 
-    public BigInteger exRecId;
+    public Long exRecId;
 
 }
